move to navigation_text
