+++ src/mars_patcher/patcher.py
@@ -13,7 +13,6 @@ from mars_patcher.locations import LocationSettings
 from mars_patcher.minimap import apply_minimap_edits
 from mars_patcher.misc_patches import (
     apply_anti_softlock_edits,
-    apply_hint_security,
     apply_pbs_without_bombs,
     apply_unexplored_map,
     change_missile_limit,
@@ -107,6 +106,7 @@ def patch(
         status_update(-1, "Writing navigation text...")
         navigation_text = NavigationText.from_json(patch_data["NavigationText"])
         navigation_text.write(rom)
+        NavigationText.apply_hint_security(rom, patch_data["NavStationLocks"])
 
     # Credits
     if "CreditsText" in patch_data:
@@ -147,8 +147,6 @@ def patch(
 
     if "MinimapEdits" in patch_data:
         apply_minimap_edits(rom, patch_data["MinimapEdits"])
-
-    apply_hint_security(rom, patch_data["NavStationLocks"])
 
     write_seed_hash(rom, patch_data["SeedHash"])
